Refactored enum types into an enums folder.
