+++ src/enums/Commands.ts
@@ -1,0 +1,15 @@
+export enum Commands {
+  'highlight' = 'twitchHighlighter.highlight',
+  'unhighlight' = 'twitchHighlighter.unhighlight',
+  'unhighlightAll' = 'twitchHighlighter.unhighlightAll',
+  'unhighlightSpecific' = 'twitchHighlighter.unhighlightSpecific',
+  'startChat' = 'twitchHighlighter.startChat',
+  'stopChat' = 'twitchHighlighter.stopChat',
+  'toggleChat' = 'twitchHighlighter.toggleChat',
+  'removeTwitchClientId' = 'twitchHighlighter.removeTwitchClientId',
+  'setTwitchToken' = 'twitchHighlighter.setTwitchToken',
+  'removeTwitchToken' = 'twitchHighlighter.removeTwitchToken',
+  'refreshTreeView' = 'twitchHighlighter.refreshTreeView',
+  'gotoHighlight' = 'twitchHighlighter.gotoHighlight',
+  'removeHighlight' = 'twitchHighlighter.removeHighlight'
+}
