+++ src/enums/InternalCommands.ts
@@ -1,0 +1,3 @@
+export enum InternalCommands {
+  'removeBannedHighlights' = 'twitchHighlighter.removeBannedHighlights'
+}
