+++ src/enums/Settings.ts
@@ -1,0 +1,13 @@
+export enum Settings {
+  'channels' = 'channels',
+  'username' = 'nickname',
+  'highlightColor' = 'highlightColor',
+  'highlightBorder' = 'highlightBorder',
+  'highlightFontColor' = 'highlightFontColor',
+  'announceBot' = 'announceBot',
+  'joinMessage' = 'joinMessage',
+  'leaveMessage' = 'leaveMessage',
+  'unhighlightOnDisconnect' = 'unhighlightOnDisconnect',
+  'showHighlightsInActivityBar' = 'showHighlightsInActivityBar',
+  'usageTip' = 'usageTip'
+}
